federation: add has_events experimental endpoint
This endpoint can be used to ask to a server if it has the requested
events in its database. This is meant to be used before sending a
transaction after a long retry interval, to avoid sending events that
have already be transmitted by another server to the destination.

This should save a bit of bandwidth at the expense of latency, but this
is not a big problem in case of exceeded long retries.